Changed withTrashed finder to use options instead
This greatly improves load times when used with a MySQL backend. For some reason, MySQL ignores indexes when using `field IS NULL OR field IS NOT NULL`. For large tables, this results in long load times (we've seen samples of 3 minutes and up). Removing this kinda strange statement will resulted in way faster queries (~5 ms).